Make the Circuit struct generic to support different Orchard variants (currently supports ZSA only)
